Converted most Integer variables to Long

+++ plugins/eu.geclipse.info/src/eu/geclipse/info/glue/GlueBatchJob.java
@@ -14,7 +14,7 @@ public class GlueBatchJob extends AbstractGlueTable implements java.io.Serializa
 
   public String GlueBatchJobGlobalID; //PK
   public GlueHost glueHost; //GlueHostUniqueID
-  public Integer LocalID;
+  public Long LocalID;
   public String Status;
   public String Name;
   public String LocalOwner;
